requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.piptools
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
assemblyai==0.33.0
    # via -r requirements.piptools
beautifulsoup4==4.12.3
    # via -r requirements.piptools
blinker==1.8.2
    # via flask
boto3==1.35.9
    # via -r requirements.piptools
botocore==1.35.9
    # via
    #   -r requirements.piptools
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
distro==1.9.0
    # via openai
flask==3.0.3
    # via
    #   -r requirements.piptools
    #   flask-cors
flask-cors==5.0.0
    # via -r requirements.piptools
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   assemblyai
    #   openai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jiter==0.5.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
openai==1.43.0
    # via -r requirements.piptools
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.piptools
pydantic==2.8.2
    # via
    #   assemblyai
    #   openai
pydantic-core==2.20.1
    # via pydantic
pytest==8.3.2
    # via -r requirements.piptools
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.piptools
pytube==15.0.0
    # via -r requirements.piptools
requests==2.32.3
    # via
    #   -r requirements.piptools
    #   youtube-transcript-api
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   assemblyai
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via
    #   -r requirements.piptools
    #   botocore
    #   requests
websockets==13.0.1
    # via assemblyai
werkzeug==3.0.4
    # via flask
youtube-transcript-api==0.6.2
    # via -r requirements.piptools
